+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - add schema info and validation to manual plugin config editor
 - show plugin schema header and footer in plugin json editor
 - hide search button when no plugins installed
+- show help on accessories when no plugins installed
 
 ### Other Changes